[ELTA-608] Add glucometer commands

+++ domain/src/main/java/com/elta/android/domain/features/devices/model/GlucometerInfo.kt
@@ -4,7 +4,7 @@ import java.util.Date
 
 data class GlucometerInfo(
     val date: Date? = null,
-    val temperature: Double? = null,
+    val temperature: Int? = null,
     val batteryLevel: Int? = null,
     val softwareVersion: Double? = null,
     val hardwareVersion: Double? = null
